# docs/api/openapi-spec.yml
...
openapi: 3.1.0
info:
  title: LogPilot Agent API
  description: 'Cloud-native, autonomous log-intelligence agent: ingestion, search, chat, RCA, reports, proactive
    failure forecasting and autonomy policy.


    Authenticate with POST /api/v1/auth/login, then send `Authorization: Bearer <token>`. Real-time events and streamed
    chat use the WebSocket at /ws?token=<jwt> (not described by OpenAPI).'
  version: 2.0.0
paths:
...
  /api/v1/projects/{project_id}/logs/upload:
    post:
      tags:
      - logs
      summary: Upload
      description: 'Streams the multipart body straight to the ingestion service (never buffered here). The 500MB
        limit is

        enforced both on Content-Length and on bytes actually received.'
      operationId: upload_api_v1_projects__project_id__logs_upload_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: environment
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Environment
      - name: default_service
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Default Service
      - name: custom_pattern
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Custom Pattern
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/logs/records:
    post:
      tags:
      - logs
      summary: Ingest Records
      operationId: ingest_records_api_v1_projects__project_id__logs_records_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordsBody'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/logs/stream:
    post:
      tags:
      - logs
      summary: Stream
      description: 'Real-time streaming ingestion: send a chunked NDJSON body (one JSON record per line).'
      operationId: stream_api_v1_projects__project_id__logs_stream_post
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: environment
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Environment
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/logs/sessions:
    get:
      tags:
      - logs
      summary: Sessions
      operationId: sessions_api_v1_projects__project_id__logs_sessions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/logs/sessions/{session_id}:
    get:
      tags:
      - logs
      summary: Session
      operationId: session_api_v1_projects__project_id__logs_sessions__session_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/logs/sessions/{session_id}/malformed:
    get:
      tags:
      - logs
      summary: Malformed
      operationId: malformed_api_v1_projects__project_id__logs_sessions__session_id__malformed_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/projects/{project_id}/clusters:
    get:
      tags:
      - health
      summary: Clusters
      operationId: clusters_api_v1_projects__project_id__clusters_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: status
        in: query
        required: false
        schema:
          type: string
          default: active
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/dedup:
    get:
      tags:
      - health
      summary: Dedup
      operationId: dedup_api_v1_projects__project_id__dedup_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/projects/{project_id}/anomalies:
    get:
      tags:
      - health
      summary: Anomalies
      operationId: anomalies_api_v1_projects__project_id__anomalies_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          default: 24
          title: Hours
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ApiRecord:
      properties:
        timestamp:
          anyOf:
          - type: string
          - type: number
          - type: integer
          - type: 'null'
          title: Timestamp
        message:
          type: string
          title: Message
        severity:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
        level:
          anyOf:
          - type: string
          - type: 'null'
          title: Level
        service:
          anyOf:
          - type: string
          - type: 'null'
          title: Service
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
        trace_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Trace Id
        environment:
          anyOf:
          - type: string
          - type: 'null'
          title: Environment
        deployment_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Deployment Version
        attributes:
          additionalProperties: true
          type: object
          title: Attributes
      type: object
      required:
      - message
      title: ApiRecord
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RecordsBody:
      properties:
        environment:
          anyOf:
          - type: string
          - type: 'null'
          title: Environment
        records:
          items:
            $ref: '#/components/schemas/ApiRecord'
          type: array
          maxItems: 10000
          minItems: 1
          title: Records
      type: object
      required:
      - records
      title: RecordsBody
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: http://localhost:8000
  description: Local (docker compose)